% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3FORA-package.R
\docType{package}
\name{M3FORA-package}
\alias{M3FORA-package}
\alias{M3FORA}
\title{Data from the M3-competition}
\source{
\url{https://github.com/svcuonghvktqs/M3FORA}.
}
\description{
The 3003 time series from the IJF-M3 competition (Makridakis and Hibon, 2000)
and M3 Data with Rolling-Origin and Interval Forecasts
}
\details{
{\tabular{ll}{ Package: \tab M3FORA\cr Type: \tab Package\cr License: \tab GPL3\cr}}
}
\author{
Andrey Davydenko, Maxim Shcherbakov and Sai Van Cuong from Volgograd State Technical University
}
\keyword{package}
