% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ts.R
\name{plot.xts}
\alias{plot.xts}
\title{Plotting M Competition data with forecasts obtained from difference forecasting methods}
\usage{
\method{plot}{xts}(x, si, method = NULL)
}
\arguments{
\item{x}{The dataset of M-competition data in  list format as M3Year, M3Quart, M3Month and M3Other}

\item{si}{Name of a time series}

\item{method}{Name or vector of names of forecasting methods}
}
\value{
\code{plot.ts} returns a dygraphs object. This function produces a time series plot
of the selected series with forecast obtained from difference forecasting methods.
}
\description{
Functions to plot a time series from the M competition data sets, showing
both the training and test sections of the series with forecasts obtained from difference forecasting methods .
}
\examples{
plot.ts(M3Year, si = "Y1")
plot.ts(M3Year, si = "Y1", method = c("HOLT", "SINGLE", "NAIVE2"))

plot.ts(M3Quart, si = "Q4")
plot.ts(M3Quart, si = "Q4", method = c("HOLT", "SINGLE", "NAIVE2"))

plot.ts(M3Month, si = "M598")
plot.ts(M3Month, si = "M598", method = c("HOLT", "SINGLE", "NAIVE2"))

plot.ts(M3Other, si = "O15")
plot.ts(M3Other, si = "O15", method = c("HOLT", "SINGLE"))
plot.ts(M3Other, si = "O15", method = c("HOLT", "SINGLE", "NAIVE2"))

}
\seealso{
\code{\link{M3Year}}, \code{\link{M3Quart}}, \code{\link{M3Month}}, \code{\link{M3Other}}, \code{\link{M3Arima}}
}
\author{
Andrey Davydenko, Maxim and Sai Van Cuong from Volgograd State Technical University
}
\keyword{dygraphs}
\keyword{objects}
